// app/shared/models/contrato-catalogo/contrato-catalogo-item.ts
import { ContratoCatalogoEstado } from '..';
import { Garantia } from '../enums/garantia';
import { SituacaoContratoCatalogoItem } from '../enums/situacao-contrato-catalogo-item';
import { TipoCatalogo } from '../enums/tipo-catalogo';
import { TipoFrete } from '../enums/tipo-frete';
import { Marca } from '../marca';
import { Produto } from '../produto';
import { Moeda } from './../enums/moeda';
import { PessoaJuridica } from './../pessoa-juridica';
import { AprovacaoContratoCatalogoItemFornecedor } from './aprovacao-contrato-catalogo-item-fornecedor';
import { ContratoCatalogoFaturamento } from './contrato-catalogo-faturamento';
import { ContratoCatalogoItemEstado } from './contrato-catalogo-item-estado';

export class ContratoCatalogoItem {
  idContratoCatalogoItem: number;
  idContratoCatalogo: number;
  idTenant: number;
  codigo: string;
  idProduto: number;
  idMarca: number;
  idFornecedor: number;
  descricao: string;
  valor: number;
  prazoEntrega: number;
  moeda: Moeda;
  garantia: Garantia;
  frete: TipoFrete;
  loteMinimo: number;
  observacao: string;
  quantidadeTotal: number;
  quantidadeSolicitada: number;
  quantidadeDisponivel: number;
  situacao: SituacaoContratoCatalogoItem;
  fornecedor: PessoaJuridica;
  produto: Produto;
  marca: Marca;
  estadosItem: Array<ContratoCatalogoItemEstado>;
  estadosAtendimento: Array<ContratoCatalogoEstado>;
  faturamento: Array<ContratoCatalogoFaturamento>;
  stringEstadosAtendimento: string;
  aprovacaoContratoCatalogoItem: AprovacaoContratoCatalogoItemFornecedor;
  justificativaReprovacao: string;
  quantidade: number;
  favorito: boolean;
  tipoCatalogo: TipoCatalogo;
  ncm: string;
  icmsAliquota: number;
  stAliquota: number;
  difalAliquota: number;
  ipiAliquota: number;
  pisAliquota: number;
  cofinsAliquota: number;
}